Update ConvertSecondsToFrames to use float for seconds
Changed the parameter type of `ConvertSecondsToFrames` from `unsigned int` to `float` for better precision. Updated the implementation to handle edge cases where seconds or deltaTime are zero, ensuring robustness.

+++ include/game_utilities.h
@@ -12,6 +12,6 @@ namespace artifact
          * @param deltaTime The number of milliseconds since the last frame (60fps = 16.666ms)
          * @return The amount of frames that should elapse over 'x' seconds
          */
-        static unsigned int ConvertSecondsToFrames(unsigned int seconds, float deltaTime);
+        static unsigned int ConvertSecondsToFrames(float seconds, float deltaTime);
     };
 } // namespace artifact
